client: Split color util methods into colors.js
