+++ src/client/javascript/utils/misc.js
@@ -223,75 +223,3 @@ export function cloneObject(o, alternative) {
     return value;
   }));
 }
-
-/////////////////////////////////////////////////////////////////////////////
-// COLORS
-/////////////////////////////////////////////////////////////////////////////
-
-/**
- * Convert HEX to RGB
- *
- * @param   {String}      hex     The hex string (with #)
- *
- * @return  {Object}              RGB in form of r, g, b
- */
-export function convertToRGB(hex) {
-  const rgb = parseInt(hex.replace('#', ''), 16);
-  const val = {};
-  val.r = (rgb & (255 << 16)) >> 16;
-  val.g = (rgb & (255 << 8)) >> 8;
-  val.b = (rgb & 255);
-  return val;
-}
-
-/**
- * Convert RGB to HEX
- *
- * @param   {Number|Object}    r         Red value or RGB object
- * @param   {Number|undefined} [g]       Green value
- * @param   {Number|undefined} [b]       Blue value
- *
- * @return  {String}              Hex string (with #)
- */
-export function convertToHEX(r, g, b) {
-  if ( typeof r === 'object' ) {
-    g = r.g;
-    b = r.b;
-    r = r.r;
-  }
-
-  if ( typeof r === 'undefined' || typeof g === 'undefined' || typeof b === 'undefined' ) {
-    throw new Error('Invalid RGB supplied to convertToHEX()');
-  }
-
-  const hex = [
-    parseInt(r, 10).toString( 16 ),
-    parseInt(g, 10).toString( 16 ),
-    parseInt(b, 10).toString( 16 )
-  ];
-
-  Object.keys(hex).forEach((i) => {
-    if ( hex[i].length === 1 ) {
-      hex[i] = '0' + hex[i];
-    }
-  });
-
-  return '#' + hex.join('').toUpperCase();
-}
-
-/**
- * Ivert HEX color
- * @link http://stackoverflow.com/a/9601429/1236086
- *
- * @param   {String}      hex     Hex string (With #)
- *
- * @return  {String}              Inverted hex (With #)
- *
- */
-export function invertHEX(hex) {
-  let color = parseInt(hex.replace('#', ''), 16);
-  color = 0xFFFFFF ^ color;
-  color = color.toString(16);
-  color = ('000000' + color).slice(-6);
-  return '#' + color;
-}
